refactor(tests): separate fixture corpus targets
