+++ QtProject/app/comparison/comparison.h
@@ -35,7 +35,6 @@ struct VideoMetadata;
 class Comparison : public QDialog
 {
     Q_OBJECT
-    friend class TestLocalVideoCorpus;
     friend class TestAutoDelete;
     friend class test_comparison;
