--- QtProject/tests/test_local_video_corpus/tst_video.cpp
@@ -1,938 +1,0 @@
-#include <QApplication>
-#include <QCoreApplication>
-#include <QElapsedTimer>
-#include <QtConcurrent/QtConcurrent>
-#include <QtTest>
-
-#include <algorithm>
-
-// add necessary includes here
-#include "../../app/prefs.h"
-#include "../../app/video.h"
-
-#include "../../app/comparison/comparison.h"
-#include "../../app/comparison/ui_comparison.h"
-#include "../../app/db.h"
-#include "../../app/mainwindow.h"
-
-#include "../test_repo_video_extraction_regression/video_extraction_test_helpers.h"
-
-#include "video_test_helpers.h"
-
-class TestLocalVideoCorpus : public QObject
-{
-    Q_OBJECT
-  public:
-    TestLocalVideoCorpus();
-    ~TestLocalVideoCorpus();
-
-    // Special Qt Test static method called before main() runs
-    // Use this to set QTEST_FUNCTION_TIMEOUT for long-running tests
-    static void initMain();
-
-  private:
-#ifdef Q_OS_WIN
-    QDir _videoDir = QDir("Y:/Videos/");
-    const QDir _thumbnailDir_nocache = QDir("Y:/Thumbnails-nocache/");
-    const QFileInfo _csvInfo_nocache =
-        QFileInfo("C:/Dev/video-simili-duplicate-cleaner/QtProject/tests/test_local_video_corpus/ressources/tests-nocache.csv");
-
-    const QDir _thumbnailDir_cached = QDir("Y:/Thumbnails-cached/");
-    const QFileInfo _csvInfo_cached =
-        QFileInfo("C:/Dev/video-simili-duplicate-cleaner/QtProject/tests/test_local_video_corpus/ressources/tests-cached.csv");
-
-    QDir _100GBvideoDir = QDir("");
-    const QDir _100GBthumbnailDir_nocache = QDir("");
-    const QFileInfo _100GBcsvInfo_nocache = QFileInfo("");
-#elif defined(Q_OS_MACOS)
-    QDir _videoDir = QDir("/Users/theophanemayaud/Dev/Videos across all formats with duplicates of all kinds/Videos");
-    const QDir _thumbnailDir_nocache =
-        QDir("/Users/theophanemayaud/Dev/Videos across all formats with duplicates of all kinds/Thumbnails-nocache/");
-    const QFileInfo _csvInfo_nocache = QFileInfo(
-        "/Users/theophanemayaud/Dev/Videos across all formats with duplicates of all kinds/tests-nocache.csv");
-
-    const QDir _thumbnailDir_cached =
-        QDir("/Users/theophanemayaud/Dev/Videos across all formats with duplicates of all kinds/Thumbnails-cached/");
-    const QFileInfo _csvInfo_cached =
-        QFileInfo("/Users/theophanemayaud/Dev/Videos across all formats with duplicates of all kinds/tests-cached.csv");
-
-    QDir _100GBvideoDir = QDir(
-        "/Volumes/4TBSSD/Video duplicates - just for checking later my video duplicate program still works/Videos/");
-    const QDir _100GBthumbnailDir_nocache = QDir("/Volumes/4TBSSD/Video duplicates - just for checking later my video "
-                                                 "duplicate program still works/Thumbnails-nocache/");
-    const QFileInfo _100GBcsvInfo_nocache = QFileInfo("/Volumes/4TBSSD/Video duplicates - just for checking later my "
-                                                      "video duplicate program still works/100GBtests-nocache.csv");
-#endif
-
-  private slots:
-    void initTestCase();
-
-    void emptyDb()
-    {
-        Prefs prefs;
-        Db::initDbAndCacheLocation(prefs);
-        Db::emptyAllDb(prefs);
-    }
-
-    // Check ref params with no cache - DATA-DRIVEN TESTS
-    void test_check_refvidparams_nocache_data() { populateRefVidParamsTestData(_videoDir, 229); };
-    void test_check_refvidparams_nocache()
-    {
-        refVidParamsTestConfig conf;
-        conf.testDescr = __FUNCTION__;
-        conf.cacheOption = Prefs::NO_CACHE;
-
-        conf.acceptSmallDurationDiff = true;
-
-        checkSingleVideoParams(conf);
-    };
-
-    void test_check_refvidparams_withcache_data() { populateRefVidParamsTestData(_videoDir, 229); };
-    void test_check_refvidparams_withcache()
-    {
-        refVidParamsTestConfig conf;
-        conf.testDescr = __FUNCTION__;
-        conf.cacheOption = Prefs::WITH_CACHE;
-        conf.acceptSmallDurationDiff = true;
-        checkSingleVideoParams(conf);
-    };
-    void test_check_refvidparams_withCacheOnly_data() { populateRefVidParamsTestData(_videoDir, 229); };
-    void test_check_refvidparams_withCacheOnly()
-    {
-        refVidParamsTestConfig conf;
-        conf.testDescr = __FUNCTION__;
-        conf.cacheOption = Prefs::CACHE_ONLY;
-        conf.acceptSmallDurationDiff = true;
-        checkSingleVideoParams(conf);
-    };
-
-    // void test_check_refvidparams_nocache_manualCompare_data(){
-    //     populateRefVidParamsTestData(_videoDir, 218);
-    // };
-    // void test_check_refvidparams_nocache_manualCompare(){
-    //     refVidParamsTestConfig conf;
-    //     conf.testDescr = __FUNCTION__;
-    //     conf.cacheOption = Prefs::NO_CACHE;
-    //     conf.acceptSmallDurationDiff = true;
-    //     conf.compareThumbsVisualConfig->manualCompareIfThumbsVisualDiff = true;
-    //     conf.compareThumbsVisualConfig->sampledManualThumbVerifInterval = 1;
-    //     checkSingleVideoParams(conf);
-    // };
-
-    // void test_check_refvidparams_nocache_manualCompare10Sampled_data(){
-    //     populateRefVidParamsTestData(_videoDir, 218);
-    // };
-    // void test_check_refvidparams_nocache_manualCompare10Sampled(){
-    //     refVidParamsTestConfig conf;
-    //     conf.testDescr = __FUNCTION__;
-    //     conf.cacheOption = Prefs::NO_CACHE;
-    //     conf.paramsCSV = _csvInfo_nocache;
-    //     conf.thumbsDir = _thumbnailDir_nocache;
-    //     conf.videoDir = _videoDir;
-    //     conf.acceptSmallDurationDiff = true;
-    //     conf.compareThumbsVisualConfig->manualCompareIfThumbsVisualDiff = true;
-    //     conf.compareThumbsVisualConfig->sampledManualThumbVerifInterval = 10;
-    //     checkSingleVideoParams(conf);
-    // };
-
-    // // Utilities to regenerate reference data - uncomment when needed
-    // void createRefVidParams_nocache() {
-    //     createRefVidParams(
-    //         _videoDir,
-    //         Prefs::NO_CACHE
-    //     );
-    // };
-    // void createRefVidParams_withcache() {
-    //     createRefVidParams(
-    //         _videoDir,
-    //         Prefs::WITH_CACHE
-    //     );
-    // };
-    // void createRefVidParams100GB_nocache() {
-    //     createRefVidParams(
-    //         _100GBvideoDir,
-    //         Prefs::NO_CACHE
-    //     );
-    // };
-
-    // Test whole app
-    void test_whole_app_nocache()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::NO_CACHE;
-        conf.ref_ms_time = 4.0 * 1000;
-        conf.nb_vids_to_find = 229;
-        conf.nb_valid_vids_to_find = 226;
-        conf.nb_matching_vids_to_find = 89;
-        runWholeAppScan(_videoDir, &conf);
-    };
-
-    void test_whole_app_cached()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::WITH_CACHE;
-        conf.ref_ms_time = 1.5 * 1000;
-        conf.nb_vids_to_find = 229;
-        conf.nb_valid_vids_to_find = 226;
-        conf.nb_matching_vids_to_find = 89;
-        runWholeAppScan(_videoDir, &conf);
-    };
-
-    void test_whole_app_cache_only()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::CACHE_ONLY;
-        conf.ref_ms_time = 0.8 * 1000;
-        conf.nb_vids_to_find = 229;
-        conf.nb_valid_vids_to_find = 226;
-        conf.nb_matching_vids_to_find = 89;
-        runWholeAppScan(_videoDir, &conf);
-    };
-
-    // 100GB tests - data-driven approach
-    void test_100GBcheck_refvidparams_nocache_data() { populateRefVidParamsTestData(_100GBvideoDir, 12506); };
-    void test_100GBcheck_refvidparams_nocache()
-    {
-        refVidParamsTestConfig conf;
-        conf.testDescr = __FUNCTION__;
-        conf.cacheOption = Prefs::NO_CACHE;
-        conf.acceptSmallDurationDiff = true;
-        checkSingleVideoParams(conf);
-    };
-
-    // void test_100GBcheck_refvidparams_withcache_data() {
-    //     populateRefVidParamsTestData(_100GBvideoDir, 12506);
-    // };
-    // void test_100GBcheck_refvidparams_withcache() {
-    //     refVidParamsTestConfig conf;
-    //     conf.testDescr = __FUNCTION__;
-    //     conf.cacheOption = Prefs::WITH_CACHE;
-    //     conf.acceptSmallDurationDiff = true;
-    //     checkSingleVideoParams(conf);
-    // };
-
-    // void test_100GBcheck_refvidparams_nocache_manualCompare_data() {
-    //     populateRefVidParamsTestData(_100GBvideoDir, 12506);
-    // };
-    // void test_100GBcheck_refvidparams_nocache_manualCompare() {
-    //     refVidParamsTestConfig conf;
-    //     conf.testDescr = __FUNCTION__;
-    //     conf.cacheOption = Prefs::NO_CACHE;
-    //     conf.acceptSmallDurationDiff = true;
-    //     conf.compareThumbsVisualConfig->manualCompareIfThumbsVisualDiff = true;
-    //     conf.compareThumbsVisualConfig->sampledManualThumbVerifInterval = 500;
-    //     checkSingleVideoParams(conf);
-    // };
-
-    void test_100GBwholeApp_nocache()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::NO_CACHE;
-        conf.ref_ms_time = 4 * 60 * 1000;
-        conf.nb_vids_to_find = 12506;
-        conf.nb_valid_vids_to_find = 12352;
-        conf.nb_matching_vids_to_find = 6552;
-        runWholeAppScan(_100GBvideoDir, &conf);
-    };
-    void test_100GBwholeApp_nocache_rotations()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::NO_CACHE;
-        conf.detectRotatedCopies = true;
-        conf.ref_ms_time = 6 * 60 * 1000;
-        conf.nb_vids_to_find = 12506;
-        conf.nb_valid_vids_to_find = 12352;
-        // Calibrated by the explicit local benchmark; pair-level precision is
-        // asserted by the labeled tracked and /Dev feature matrices instead.
-        conf.nb_matching_vids_to_find = 6556;
-        runWholeAppScan(_100GBvideoDir, &conf);
-    };
-    void test_100GBwholeApp_cached()
-    {
-        wholeAppTestConfig conf;
-        conf.cacheOption = Prefs::WITH_CACHE;
-        conf.ref_ms_time = 50 * 1000;
-        conf.nb_vids_to_find = 12506;
-        conf.nb_valid_vids_to_find = 12351;
-        conf.nb_matching_vids_to_find = 6543;
-        runWholeAppScan(_100GBvideoDir, &conf);
-    };
-
-    // Runs at end of all test run (not once per test)
-    void cleanupTestCase() {};
-
-  private:
-    class wholeAppTestConfig
-    {
-      public:
-        Prefs::USE_CACHE_OPTION cacheOption;
-
-        // inside the app it default to 100, but for tests it's could be interesting if lower
-        uint videoSimilarityThreshold = 100;
-        bool detectRotatedCopies = false;
-
-        /* Small test set
-         *  - No cache
-         *      -- 207: before remove big files
-         *      -- 200
-         *      -- 209: after adding 9 new videos including gps, oct 2025
-         *      -- 218: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 229: after adding 11 matching feature fixtures, aug 2026
-         *  - Cached
-         *      -- 207: before remove big files
-         *      -- 200
-         *      -- 209: after adding 9 new videos including gps, oct 2025
-         *      -- 218: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 229: after adding 11 matching feature fixtures, aug 2026
-         * 100GB test set
-         *  - No cache
-         *      -- 12 505
-         *      -- 12 506: matching improvements audit, aug 2026
-         *  - Cached
-         *      -- 12 505
-         *      -- 12 506: matching improvements audit, aug 2026
-         */
-        int nb_vids_to_find;
-
-        /* Small test set
-         *  - No cache
-         *      -- 204: before remove big file tests
-         *      -- 197
-         *      -- 205: after adding 9 new videos including gps, oct 2025
-         *      -- 214: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 224: after adding 11 matching feature fixtures, before matching improvements, aug 2026
-         *      -- 226: after matching improvements recover two feature fixtures, aug 2026
-         *  - Cached
-         *      -- 204: before remove big file tests
-         *      -- 197
-         *      -- 205: after adding 9 new videos including gps, oct 2025
-         *      -- 214: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 224: after adding 11 matching feature fixtures, before matching improvements, aug 2026
-         *      -- 226: after matching improvements recover two feature fixtures, aug 2026
-         *  - Cache only
-         *      -- 204: before remove big file tests
-         *      -- 197
-         *      -- 205: after adding 9 new videos including gps, oct 2025
-         * 100GB test set
-         *  - No cache
-         *      -- 12 328: after redo of caching
-         *      -- 12 330: arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 12 331: November 2025 update, before delete custom threadpool
-         *      -- 12 352: matching improvements, aug 2026
-         *  - Cached
-         *      -- 12 330: after redo of caching
-         *      -- 12 330: arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 12 331: November 2025 update, before delete custom threadpool
-         *      -- 12 351: warm-cache matching improvements, aug 2026
-         */
-        int nb_valid_vids_to_find;
-
-        /* Small test set
-         *  - No cache
-         *      -- 71 (before some changes that made it go down 1)
-         *      -- 70 noticed as of 2025 oct.
-         *      -- 75: after adding 9 new videos including gps, oct 2025
-         *      -- 84: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 85: after adding 11 matching feature fixtures, before matching improvements, aug 2026
-         *      -- 89: after matching improvements, aug 2026
-         *  - Cached
-         *      -- 72 (before some changes that made it go down 1)
-         *      -- 71 noticed as of 2025 oct.
-         *      -- 75: after adding 9 new videos including gps, oct 2025
-         *      -- 84: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 85: after adding 11 matching feature fixtures, before matching improvements, aug 2026
-         *      -- 89: after matching improvements, aug 2026
-         *  - Cache only
-         *      -- 74: before remove big file tests
-         *      -- 72 (before some changes that made it go down 1)
-         *      -- 71 noticed as of 2025 oct.
-         *      -- 75: after adding 9 new videos including gps, oct 2025
-         *      -- 84: after adding 9 more videos in apple photos library, nov 2025
-         *      -- 85: after adding 11 matching feature fixtures, before matching improvements, aug 2026
-         *      -- 89: after matching improvements, aug 2026
-         * 100GB test set
-         *  - No cache
-         *      -- 6 626: mix lib&exec metadata, exec captures
-         *      -- 6 562: lib(only) metadata, lib(only) captures
-         *      -- 6 558 (97.1GB): arm but intel build
-         *      -- 6 568 (97.2 GB): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 6 535 (97.0 GB): November 2025 update, before delete custom threadpool
-         *      -- 6 552 off / 6 556 rotation on: matching improvements, aug 2026
-         *  - Cached
-         *      -- 6 626: mix lib&exec metadata, exec captures
-         *      -- 6 553: lib(only) metadata, lib(only) captures
-         *      -- 6 550 (97.1GB): after redo of caching
-         *      -- 6 555 (97.1 GB): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 6 527 (97.0 GB): November 2025 update, before delete custom threadpool
-         *      -- 6 543: warm-cache matching improvements, aug 2026
-         */
-        int nb_matching_vids_to_find;
-
-        /* Small test set
-         *  - No cache
-         *      -- 13.5 sec: windows intel on intel (before remove big file tests)
-         *      -- 13 sec: macOS intel on intel (before remove big file tests)
-         *      -- 6 sec: macOS arm on M1
-         *      -- 3.122 sec (3.37, 2.995,...): arm m3 Pro with arm build & arm lib - 2024 oct
-         *      -- 4.5 sec (3.996s, 4.3s, 4.13s...): after adding 9 new videos including gps, m3 MBP oct 2025
-         *      -- 4.0 sec (3.701s, 3.871s, 3.803s): arm m3 pro - nov 2025 - after removing custom threadpool
-         *  - Cached
-         *      -- 2.75 sec: windows intel on intel (before remove big file tests 207)
-         *      -- 3 sec: macOS intel on intel (before remove big file tests 207)
-         *      -- 1 sec: macOS arm on M1
-         *      -- 650 ms (0.572, 0.570,...): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 2.0 sec (1.136s, 1.109s, 1.125s...): after adding 9 new videos including gps, m3 MBP oct 2025
-         *      -- 1.5 sec (1.115s, 1.153s, 1.141s) arm m3 pro - nov 2025 - after removing custom threadpool (bimodal: ~1.1s or ~1.9s due to system scheduling)
-         *  - Cache only
-         *      -- 3.203 sec: macos intel on intel (before remove big file tests 207)
-         *      -- <1 sec: macOS arm on M1
-         *      -- <1 sec (0.566, 0.538,...): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 1.0 sec (0.571s, 0.567s, 0.568s...): after adding 9 new videos including gps, m3 MBP oct 2025
-         *      -- 0.8 sec (0.694s, 0.642s, 0.701s): arm m3 pro - nov 2025 - after removing custom threadpool
-         * 100GB test set
-         *  - No cache
-         *      -- 36min: mix lib&exec metadata, exec captures
-         *      -- 30min: lib(only) metadata, exec captures
-         *      -- 17min: lib(only) metadata, lib(only) captures
-         *      -- 17min: lib(only) metadata, lib(only) captures
-         *      -- 6min 30s (418s, ...): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 5min (250s, 263s, 287s...): arm m3 Pro with arm build & arm lib - 2024 dec. after ordered lists reworking, and 2025 march with custom task pool
-         *      -- 5min (273.736s, 255.859s): arm m3 Pro - November 2025 update, before delete custom threadpool
-         *      -- 4min (237.977s, 236.576s, 234.792s): arm m3 pro - nov 2025 - after delete custom threadpool
-         *  - Cached
-         *      -- 17min: mix lib&exec metadata, exec captures
-         *      -- 6min: lib(only) metadata, exec captures
-         *      -- 6min: lib(only) metadata, lib(only) captures
-         *      -- 2min 33s: arm m3 Pro with arm build & arm lib - 2024 oct.
-         *      -- 50s (44s, 46s, 51s, 43s,...): arm m3 Pro with arm build & arm lib - 2024 dec. after ordered lists reworking, and 2025 march with custom task pool
-         *      -- 50s (42s, 44s, 42.262s): arm m3 Pro - November 2025 update, before delete custom threadpool
-         *      -- 50s (45.397s, 43.660s, 44.903s): arm m3 pro - nov 2025 - after delete custom threadpool
-         *  */
-        qint64 ref_ms_time;
-    };
-
-    class refVidParamsTestConfig
-    {
-      private:
-        class visualCompareTestConfig;
-
-      public:
-        refVidParamsTestConfig() {}
-
-        QString testDescr;
-        QFileInfo paramsCSV;
-        QDir videoDir;
-        QDir thumbsDir;
-
-        /* Thumbnails are generated without cache : must run test without cache (clean all before)
-         * Also sometimes, for some unknown reason, thumbnails don't come out the same.
-         * But if you re-run tests a few times, it should get fixed
-         * (or check visually with compareThumbsVisualConfig) */
-        Prefs::USE_CACHE_OPTION cacheOption;
-
-        /* Small test set
-         *  - No cache
-         *      -- 35s: macOS intel on intel (before remove big file tests)
-         *      -- 36s: windows intel on intel (before remove big file tests)
-         *      -- 17s: macOS arm m3 Pro with arm build & arm lib - 2024 oct.
-         *  - Cached
-         *      -- 8s: macOS intel on intel (before remove big file tests)
-         *      -- 9s: windows intel on intel (before remove big file tests)
-         *      -- 2.5s (2'494ms, 2'606ms, ...): macOS arm m3 Pro with arm build & arm lib - 2024 oct.
-         * 100GB test set
-         *  - No cache
-         *      -- 37min: library(only) metadata, exec captures
-         *      -- 48min: intel i5 lib(only) metadata, lib(only) captures
-         *      -- 26min (1568s, 1594s, ...): arm m3 Pro with arm build & arm lib - 2024 oct.
-         *  - Cached
-         *      -- 39min: mix lib&exec metadata, exec captures
-         *      -- 12min: library(only) metadata, exec captures
-         *      -- 12min: lib(only) metadata, lib(only) captures
-         *      -- 3min 27s (205s, ...: arm m3 Pro with arm build & arm lib (no thumb check) - 2024 oct. */
-        qint64 refDuration_ms;
-
-        // inside the app it default to 100, but for tests it's could be interesting if lower
-        uint videoSimilarityThreshold = 100;
-
-        // if not set, thumbnail visual details will not be checked
-        visualCompareTestConfig* compareThumbsVisualConfig = new visualCompareTestConfig();
-
-        // When moving over to library, audio metadata sometimes changes but when manually checked, is actually identical
-        // Disable following define to skip testing audio comparison
-        bool compareAudio = true;
-
-        bool compareModifiedDates = true;
-        bool acceptSmallDurationDiff = false;
-
-      private:
-        class visualCompareTestConfig
-        {
-          public:
-            visualCompareTestConfig() {}
-
-            bool manualCompareIfThumbsVisualDiff = false;
-
-            // Sometimes hashes go crazy, so we can manually disable them to see if other problems exist
-            bool compareThumbHashes = true;
-
-            uint sampledManualThumbVerifInterval = 1;
-        };
-    };
-
-    /*
-     * Runs the whole app for the given folder, by default using the cache but not checking any details
-     * If provided a conf, it will make sure to test the chosen cache option, and verify the number of elements found
-    */
-    void runWholeAppScan(QDir videoDir, wholeAppTestConfig* conf = nullptr);
-
-    // Helper to populate test data for data-driven tests
-    void populateRefVidParamsTestData(const QDir videoDir, const int expectedVideoCount);
-
-    // Helper to perform single video test (called for each test data row)
-    void checkSingleVideoParams(const refVidParamsTestConfig conf);
-
-    static void compareVideoParamToVideoAndUpdateThumbIfVisuallyIdentifcal(
-        const QByteArray ref_thumbnail, const VideoParam videoParam, const Video* vid,
-        const refVidParamsTestConfig conf = refVidParamsTestConfig());
-
-    void createRefVidParams(QDir videoDir, Prefs::USE_CACHE_OPTION cacheOption);
-};
-
-TestLocalVideoCorpus::TestLocalVideoCorpus() {}
-
-TestLocalVideoCorpus::~TestLocalVideoCorpus() {}
-
-void TestLocalVideoCorpus::initMain()
-{
-    // Set timeout to 2 hours (7200000ms) for long-running 100GB tests
-    // Default is 300000ms (5 minutes)
-    qputenv("QTEST_FUNCTION_TIMEOUT", "7200000");
-}
-
-// Runs once before test run (not once per test)
-void TestLocalVideoCorpus::initTestCase()
-{
-    qSetMessagePattern("%{file}(%{line}) %{function}: %{message}");
-    Prefs().resetSettings();
-    qDebug() << "Cleared settings";
-}
-
-// ----------------------------------------------------------------------------------
-// ---------------------------- START : helper testing functions ----------------------
-void TestLocalVideoCorpus::runWholeAppScan(QDir videoDir, wholeAppTestConfig* conf)
-{
-    QVERIFY(videoDir.exists());
-
-    if (conf != nullptr && conf->cacheOption != Prefs::NO_CACHE)
-        runWholeAppScan(videoDir); // run once to make sure all is cached
-
-    MainWindow w = MainWindow();
-    w.show();
-
-    if (conf != nullptr) {
-        w.on_thresholdSlider_valueChanged(conf->videoSimilarityThreshold);
-        w.ui->detectRotatedCopiesCheckbox->setChecked(conf->detectRotatedCopies);
-
-        switch (conf->cacheOption) {
-        case Prefs::CACHE_ONLY:
-            w.ui->radio_UseCacheOnly->click();
-            break;
-        case Prefs::WITH_CACHE:
-            w.ui->radio_UseCacheYes->click();
-            break;
-        case Prefs::NO_CACHE:
-            w.ui->radio_UseCacheNo->click();
-        default:
-            break;
-        }
-    }
-    else
-        w.ui->radio_UseCacheYes->click();
-
-    QElapsedTimer timer;
-    timer.start();
-
-    w.findVideos(videoDir);
-    w.processVideos();
-
-    qDebug() << "runWholeAppScan found " << w._everyVideo.count() << " files of which " << w._videoList.count()
-             << " valid ones";
-    qDebug() << "runWholeAppScan before match report took" << timer.elapsed() / 1000 << "." << timer.elapsed() % 1000
-             << " secs";
-
-    Comparison comp(w._videoList, w._prefs, w.geometry());
-    int matchingVideoNb = comp.reportMatchingVideos();
-    qDebug() << "runWholeAppScan found " << matchingVideoNb << " matching vids";
-
-    qDebug() << "runWholeAppScan took" << timer.elapsed() / 1000 << "." << timer.elapsed() % 1000 << " secs";
-
-    if (conf != nullptr) {
-        QVERIFY2(w._everyVideo.count() == conf->nb_vids_to_find,
-                 QString("runWholeAppScan found %1 files but should be %2")
-                     .arg(w._everyVideo.count())
-                     .arg(conf->nb_vids_to_find)
-                     .toStdString()
-                     .c_str());
-        QVERIFY2(w._videoList.count() == conf->nb_valid_vids_to_find,
-                 QString("runWholeAppScan found %1 valid files but should be %2")
-                     .arg(w._videoList.count())
-                     .arg(conf->nb_valid_vids_to_find)
-                     .toStdString()
-                     .c_str());
-        QVERIFY2(matchingVideoNb == conf->nb_matching_vids_to_find,
-                 QString("runWholeAppScan found %1 matching vids but should be %2")
-                     .arg(matchingVideoNb)
-                     .arg(conf->nb_matching_vids_to_find)
-                     .toStdString()
-                     .c_str());
-        QVERIFY2(timer.elapsed() < conf->ref_ms_time,
-                 QString("runWholeAppScan took : %1.%2s but should be below %3.%4s")
-                     .arg(timer.elapsed() / 1000)
-                     .arg(timer.elapsed() % 1000)
-                     .arg(conf->ref_ms_time / 1000)
-                     .arg(conf->ref_ms_time % 1000)
-                     .toStdString()
-                     .c_str());
-    }
-
-    w.close();
-    QCoreApplication::processEvents();
-}
-
-// Helper to populate test data for data-driven tests
-void TestLocalVideoCorpus::populateRefVidParamsTestData(const QDir videoDir, const int expectedVideoCount)
-{
-    emptyDb();
-    QTest::addColumn<QString>("videoPath");
-
-    QVERIFY(videoDir.exists());
-
-    // Load extensions
-    MainWindow w;
-    w.loadExtensions();
-    QVERIFY(!w._extensionList.isEmpty());
-    w.close();
-
-    // Find all videos in directory
-    QDir scanDir = videoDir;
-    scanDir.setNameFilters(w._extensionList);
-    QDirIterator iter(scanDir, QDirIterator::Subdirectories);
-
-    QStringList videoPaths;
-    while (iter.hasNext()) {
-        QString videoPath = iter.next();
-        QFileInfo videoInfo(videoPath);
-
-        // Skip if it's a directory (shouldn't happen with nameFilters but be safe)
-        if (videoInfo.isDir())
-            continue;
-
-        videoPaths.append(videoPath);
-    }
-
-    // Verify we found the expected number of videos
-    QVERIFY2(videoPaths.count() == expectedVideoCount, QString("Expected %1 videos but found %2 in %3")
-                                                           .arg(expectedVideoCount)
-                                                           .arg(videoPaths.count())
-                                                           .arg(videoDir.path())
-                                                           .toUtf8());
-
-    // Add each video as a test row
-    foreach (const QString& videoPath, videoPaths) {
-        QFileInfo videoInfo(videoPath);
-        // Use relative path from video directory to ensure unique test row names
-        QString relativePath = videoDir.relativeFilePath(videoPath);
-        QTest::newRow(relativePath.toStdString().c_str()) << videoPath;
-    }
-}
-
-// Helper to perform single video test (called for each test data row)
-void TestLocalVideoCorpus::checkSingleVideoParams(const refVidParamsTestConfig conf)
-{
-    QFETCH(QString, videoPath);
-
-    QFileInfo videoInfo(videoPath);
-    QVERIFY2(videoInfo.exists(), videoPath.toUtf8());
-
-    // Determine suffix based on cache option
-    QString suffix = (conf.cacheOption == Prefs::NO_CACHE) ? "nocache" : "withcache";
-
-    // Load reference metadata from individual file (video.mp4.nocache.txt or video.mp4.withcache.txt)
-    QString metadataPath = videoPath + "." + suffix + ".txt";
-    QFileInfo metadataInfo(metadataPath);
-    QVERIFY2(metadataInfo.exists(), QString("Metadata file not found: %1").arg(metadataPath).toUtf8());
-
-    VideoParam videoParam;
-    VideoExtractionTestHelpers::loadMetadataFromFile(metadataPath, videoParam);
-
-    // Load reference thumbnail from individual file (video.mp4.nocache.jpg or video.mp4.withcache.jpg)
-    QString thumbnailPath = videoPath + "." + suffix + ".jpg";
-    QByteArray ref_thumbnail = VideoExtractionTestHelpers::loadThumbnailFromFile(thumbnailPath);
-
-    Prefs prefs;
-    prefs.useCacheOption(conf.cacheOption);
-
-    // run scan once to populate cache
-    if (conf.cacheOption != Prefs::NO_CACHE) {
-        QVERIFY2(Db::initDbAndCacheLocation(prefs), "Failed to initialize cache");
-        auto prefsWithCache = prefs;
-        prefsWithCache.useCacheOption(Prefs::WITH_CACHE); // Need to force generate cache even if CACHE_ONLY is set
-        Video* vid = new Video(prefsWithCache, videoPath);
-        auto res = vid->process();
-        delete vid;
-    }
-
-    // Process video with current settings
-    Video* vid = new Video(prefs, videoPath);
-    auto res = vid->process();
-
-    // Compare
-    compareVideoParamToVideoAndUpdateThumbIfVisuallyIdentifcal(ref_thumbnail, videoParam, vid, conf);
-
-    delete vid;
-}
-
-void TestLocalVideoCorpus::compareVideoParamToVideoAndUpdateThumbIfVisuallyIdentifcal(
-    const QByteArray ref_thumbnail, const VideoParam videoParam, const Video* vid, const refVidParamsTestConfig conf)
-{
-    const QString forVid = QString("For %1").arg(videoParam.thumbnailInfo.absoluteFilePath());
-
-    bool manuallyAccepted = false, pHashAccepted = false, ssimAccepted = false;
-
-    // pHash thumbnail comparison with original hashes computed and saved as metadata
-    // pHash works by comparing number of equal bits between two hashes
-    if (conf.compareThumbsVisualConfig->compareThumbHashes) {
-        int distance1 = 64, distance2 = 64;
-        uint64_t differentBits1 =
-            videoParam.hash1 ^ vid->fingerprint(0).phash; //XOR to value (only ones for differing bits)
-        uint64_t differentBits2 = videoParam.hash2 ^ vid->fingerprint(1).phash;
-        while (differentBits1) {
-            differentBits1 &= differentBits1 - 1; //count number of bits of value
-            distance1--;
-        }
-        while (differentBits2) {
-            differentBits2 &= differentBits2 - 1; //count number of bits of value
-            distance2--;
-        }
-        // TODO add min pHash distance as test parameter
-        const int minDistance = 60; // x same bits out of 64
-
-        if (distance1 >= minDistance && distance2 >= minDistance)
-            pHashAccepted = true;
-        else
-            qWarning()
-                << QString("hash1 distance=%1/64 hash2 distance=%2/64, expected at least %3 bits to be the same - %4")
-                       .arg(distance1)
-                       .arg(distance2)
-                       .arg(minDistance)
-                       .arg(forVid)
-                       .toUtf8();
-    }
-
-    // SSIM thumbnail comparison
-    auto ssim = VideoExtractionTestHelpers::compareThumbnails(ref_thumbnail, vid->thumbnail);
-    // TODO add ssim tolerance as test parameter
-    const auto ssimThreshold = 0.90;
-    if (ssim < ssimThreshold)
-        qWarning() << QString("got ssim %1 pct expected at least %2\% - %3")
-                          .arg(ssim * 100)
-                          .arg(ssimThreshold * 100)
-                          .arg(forVid)
-                          .toUtf8();
-    else
-        ssimAccepted = true;
-
-    QVERIFY2(videoParam.size == vid->size,
-             QString("ref size=%1 new size=%2 - %3").arg(videoParam.size).arg(vid->size).arg(forVid).toUtf8());
-
-    if (conf.compareModifiedDates) {
-        const auto ref_modified = videoParam.modified.toString(VideoParam::timeformat());
-        const auto new_modified = vid->modified.toString(VideoParam::timeformat());
-        QVERIFY2(ref_modified == new_modified,
-                 QString("Date diff : ref modified=%1 new modified=%2").arg(ref_modified).arg(new_modified).toUtf8());
-    }
-    uint acceptedDurationDiff = 0;
-    // TODO add duration tolerance size ie 50ms or 1% of ref as test parameter
-    if (conf.acceptSmallDurationDiff)
-        acceptedDurationDiff = std::max(50, int(0.01 * videoParam.duration)); // biggest of 50 ms or 1% of ref duration
-    QVERIFY2(abs(videoParam.duration - vid->duration) <= acceptedDurationDiff,
-             QString("ref duration=%1 new duration=%2, max accepted diff %3 - %4")
-                 .arg(videoParam.duration)
-                 .arg(vid->duration)
-                 .arg(acceptedDurationDiff)
-                 .arg(forVid)
-                 .toUtf8());
-
-    QVERIFY2(
-        abs(videoParam.bitrate - vid->bitrate) <= 50,
-        QString("ref bitrate=%1 new bitrate=%2 - %3").arg(videoParam.bitrate).arg(vid->bitrate).arg(forVid).toUtf8());
-    QVERIFY2(
-        videoParam.framerate == vid->framerate,
-        QString("framerate ref=%1 new=%2 - %3").arg(videoParam.framerate).arg(vid->framerate).arg(forVid).toUtf8());
-    QVERIFY2(videoParam.codec == vid->codec,
-             QString("codec ref=%1 new=%2 - %3").arg(videoParam.codec).arg(vid->codec).arg(forVid).toUtf8());
-    if (conf.compareAudio)
-        QVERIFY2(videoParam.audio == vid->audio,
-                 QString("audio ref=%1 new=%2 - %3").arg(videoParam.audio).arg(vid->audio).arg(forVid).toUtf8());
-    // Sometimes ffmpeg swaps detected width and height but that's ok as long as thumbnails are the same
-    auto width = videoParam.width, height = videoParam.height;
-    if (videoParam.width == vid->height && videoParam.height == vid->width) {
-        width = videoParam.height;
-        height = videoParam.width;
-    }
-    QVERIFY2(width == vid->width, QString("width x height ref=%1x%2 new=%3x%4 - %5")
-                                      .arg(width)
-                                      .arg(height)
-                                      .arg(vid->width)
-                                      .arg(vid->height)
-                                      .arg(forVid)
-                                      .toUtf8());
-    QVERIFY2(height == vid->height, QString("width x height ref=%1x%2 new=%3x%4 - %5")
-                                        .arg(width)
-                                        .arg(height)
-                                        .arg(vid->width)
-                                        .arg(vid->height)
-                                        .arg(forVid)
-                                        .toUtf8());
-
-    QVERIFY2(ref_thumbnail.isNull() == vid->thumbnail.isNull(),
-             QString("Ref thumb empty=%1 but new thumb empty=%2 - %3")
-                 .arg(ref_thumbnail.isNull())
-                 .arg(vid->thumbnail.isNull())
-                 .arg(forVid)
-                 .toUtf8());
-
-    if (conf.compareThumbsVisualConfig == nullptr)
-        return;
-
-    if (ref_thumbnail.isEmpty() && vid->thumbnail.isEmpty())
-        return;
-
-    if (!ssimAccepted || !pHashAccepted) {
-        if (!conf.compareThumbsVisualConfig->manualCompareIfThumbsVisualDiff)
-            QFAIL(QString("Thumbnails not exactly identical, pHashAccepted=%1, ssim Accepted=%2 - %3")
-                      .arg(pHashAccepted)
-                      .arg(ssimAccepted)
-                      .arg(forVid)
-                      .toUtf8());
-        else if (TestHelpers::doThumbnailsLookSameWindow(
-                     ref_thumbnail, vid->thumbnail,
-                     QString("Thumbnail %1").arg(videoParam.thumbnailInfo.absoluteFilePath())))
-            manuallyAccepted = true;
-        else
-            QFAIL(QString("Thumbnails not exactly identical and manually rejected - %1").arg(forVid).toUtf8());
-    }
-
-    // Update saved thumbnail if manually accepted
-    if (manuallyAccepted) {
-        QFile thumbFile(videoParam.thumbnailInfo.absoluteFilePath());
-        if (!thumbFile.open(QIODevice::WriteOnly))
-            qWarning() << "Failed to open thumbnail file for writing:" << videoParam.thumbnailInfo.absoluteFilePath();
-        else
-            thumbFile.write(vid->thumbnail);
-        thumbFile.close();
-    }
-}
-
-// Method to generate .txt, .jpg and ref metadata and thumbnail files
-void TestLocalVideoCorpus::createRefVidParams(QDir videoDir, Prefs::USE_CACHE_OPTION cacheOption)
-{
-    emptyDb();
-    QElapsedTimer timer;
-    timer.start();
-
-    if (cacheOption != Prefs::NO_CACHE)
-        runWholeAppScan(videoDir);
-
-    QVERIFY(!videoDir.isEmpty());
-    MainWindow w;
-    w.loadExtensions();
-    QVERIFY(!w._extensionList.isEmpty());
-    w.close();
-
-    // Determine suffix based on cache option
-    QString suffix = (cacheOption == Prefs::NO_CACHE) ? "nocache" : "withcache";
-    qDebug() << "====== Generating" << suffix << "reference data ======";
-
-    videoDir.setNameFilters(w._extensionList);
-    QDirIterator iter(videoDir, QDirIterator::Subdirectories);
-
-    int processedCount = 0;
-    int errorCount = 0;
-
-    Prefs prefs;
-    prefs.useCacheOption(cacheOption);
-    QVERIFY2(Db::initDbAndCacheLocation(prefs), "Failed to initialize cache");
-
-    while (iter.hasNext()) {
-        const QFile vidFile(iter.next());
-        const QFileInfo vidInfo = QFileInfo(vidFile);
-
-        if (vidInfo.isDir())
-            continue;
-
-        QString videoPath = vidInfo.absoluteFilePath();
-        QString metadataPath = videoPath + "." + suffix + ".txt";
-        QString thumbPath = videoPath + "." + suffix + ".jpg";
-
-        // Process video
-        Video* vid = new Video(prefs, videoPath);
-        auto result = vid->process();
-
-        // Create VideoParam
-        VideoParam videoParam;
-        videoParam.videoInfo = vidInfo;
-        videoParam.size = vid->size;
-        videoParam.modified = vid->modified;
-        videoParam.duration = vid->duration;
-        videoParam.bitrate = vid->bitrate;
-        videoParam.framerate = vid->framerate;
-        videoParam.codec = vid->codec;
-        videoParam.audio = vid->audio;
-        videoParam.width = vid->width;
-        videoParam.height = vid->height;
-        videoParam.hash1 = vid->fingerprint(0).phash;
-        videoParam.hash2 = vid->fingerprint(1).phash;
-        videoParam.thumbnail = vid->thumbnail;
-
-        // Remove old files if they exist
-        if (QFileInfo::exists(metadataPath))
-            QFile::remove(metadataPath);
-        if (QFileInfo::exists(thumbPath))
-            QFile::remove(thumbPath);
-
-        // Save metadata
-        if (!VideoExtractionTestHelpers::saveMetadataToFile(videoParam, metadataPath, videoDir)) {
-            qWarning() << "Failed to save metadata for:" << videoPath;
-            errorCount++;
-            delete vid;
-            continue;
-        }
-
-        // Save thumbnail
-        if (!VideoExtractionTestHelpers::saveThumbnail(videoParam.thumbnail, thumbPath)) {
-            qWarning() << "Failed to save thumbnail for:" << videoPath;
-            errorCount++;
-            delete vid;
-            continue;
-        }
-
-        processedCount++;
-        if (processedCount % 10 == 0)
-            qDebug() << "Progress:" << processedCount << "videos processed...";
-
-        delete vid;
-    }
-
-    qDebug() << "TIMER: createRefVidParams took" << timer.elapsed() / 1000 << "." << timer.elapsed() % 1000 << " secs";
-    qDebug() << "Successfully processed:" << processedCount << "videos";
-    qDebug() << "Errors:" << errorCount << "videos";
-    QVERIFY2(errorCount == 0, QString("Some videos had errors: %1").arg(errorCount).toUtf8());
-}
-
-// ---------------------------- END : helper testing functions ------------------------
-// ----------------------------------------------------------------------------------
-
-QTEST_MAIN(TestLocalVideoCorpus)
-
-#include "tst_video.moc"
