+++ QtProject/tests/test_repo_video_extraction_regression/tst_video_extraction_regression.cpp
@@ -5,7 +5,7 @@
 #include "../../app/db.h"
 #include "../../app/prefs.h"
 #include "../../app/video.h"
-#include "video_extraction_test_helpers.h"
+#include "../video_extraction_test_helpers.h"
 
 class TestVideoExtractionRegression : public QObject
 {
